Add support for merchants logo and banner
Add a new `logo` and `banner` attachment to `Merchant` model.

In admin side, `#edit` and `#update` actions have been added to allow
uploading files through form.

In frontend, the logo is displayed if it has been attached, otherwhise
merchants initials are displayed.

To avoid duplicating attachable code between `Directory` and `Merchant` models,
a new `WithLogoAndBanner` concern has been created. Indeed, both models allowed a
logo and a banner to be attached to a record with similar logic for both.