Refactor map filter form to be more efficient
Current form version has some drawbacks with keyboard navigation and accessibility.
Moreover submit code is complicated with messy conditions to check if form should be
submitted or not according to press keys.

This new version clean the form configuration as well as the Stimulus controller with
a more accurate keyboard press key detection which make it accessible as expected.

Other related changes:
- Move back map page title below content for table and grid display (more space for
real content)
- Fix map referer url not being cleaned before saved in session (empty GET params)
- Fix some CSS styling issues on form input and outline properly merchant cards
- Fix some HTML warning with atm input that was not wrapped by a `<form>` tag