Handle `nil` value for merchants `#ask_kyc`
Do not assume anymore that a merchant does not ask for KYC when `payment:kyc` property
is not set. Instead set it to `nil` in addition to `true` and `false` values.

That means the boolean field becomes a volontary three states boolean and `nil` value
must be considered independantly.

Refactoring touches `MercantData` class, `MerchantProposal` form and views.
A migration removes the `not null` constraint on `ask_kyc` column.