feat: remember last used locale in session
Currently, URL own the thruth on which language a page should be rendered.
By default, if no locale params is present in the URL, it uses the default
locale (english).

With this change, last used locale is now stored in session allowing a page
without the locale params to be smartly redirected to the user last used locale
in the website.

This feature would be very helpful for short urls such as `/asdb` or `/map`.

Be careful, as more redirections are involved in the process, some pages might
take longer to load.