feat: add announcement resource to display news publicly
This new feature brings an `Announcement` model that is manageable from admin
side for any super admin.

It is possible to configure, for each announcement, the locale it should be
shown, the title, description, picture, link to visit and (un)publication dates.

Announcements are following a color code depending of the news mode (info, success,
warning, alert, default) to quickly identify the scope.

Minor CSS adjustement have been made to enhance the announcement experience.