feat: store model string and text columns in database for each locales
Introduce the `mobility` gem to store dynamically column values that need
to be translated in other languages.

Admin forms have been updated to display a tabs component with tab for each
locales.

Translatable columns need a validation for the default locale as it is used
as fallback content for other locales.