feat: add help modal that display keyboard shortcuts
A new modal displays the existing keyboard shortcut combination in the website.
Help button is only show in the footer on desktop as it has no meaning to be displayed
on mobile devices that does not have keyboard.

Keyboard shortcut that open the modal is `Shift` + `,` on AZERTY
or `?` on QUERTY.